Text-to-SQL:tuganbasaran_zehrasagin/requirements.txt
pandas 
SQLAlchemy 
python-dotenv
matplotlib
streamlit
google-generativeai
python-docx
pymysql
cryptography
statsmodels
scikit-learn
plotly